+++ lib/items/tool-bar-button-view.js
