+++ lib/items/tool-bar-item.js
