+++ lib/items/tool-bar-spacer-view.js
