+++ lib/tool-bar-manager.js
@@ -1,6 +1,6 @@
-const { ToolBarItem } = require('./tool-bar-item');
-const { ToolBarButtonView } = require('./tool-bar-button-view');
-const { ToolBarSpacerView } = require('./tool-bar-spacer-view');
+const { ToolBarItem } = require('./items/tool-bar-item');
+const { ToolBarButtonView } = require('./items/tool-bar-button-view');
+const { ToolBarSpacerView } = require('./items/tool-bar-spacer-view');
 
 module.exports = class ToolBarManager {
   constructor (group, toolBarView, touchBarManager) {
